refactor: delete unused files, put common files into common_util
