+++ src/common_util/fk.py
@@ -1,7 +1,7 @@
 import numpy as np
 import py_ecc.bn128 as b
 from src.common_util.curve import Scalar
-from src.cq.fft import fft, ec_fft, ec_ifft, is_power_of_two
+from src.common_util.fft import fft, ec_fft, ec_ifft, is_power_of_two
 
 # https://eprint.iacr.org/2023/033
 def fk(coeffs, powers_of_x):
